Move OpenCode report extraction to owned module
